simplify range loops (#2)
* add go.mod, update imports and format source code using goimports

This PR adds a go.mod file to the repository, thus tracking
dependencies of the project. The following command was run to
create the go.mod file:

	go mod init github.com/matthewrkarlsen/xcs-in-go

The PR also updates package import paths from local and relative to
fully qualified, e.g. "../mli" -> "github.com/matthewrkarlsen/xcs-in-go/pkg/mli"

Finally, the PR formats the source code and groups imports using
the goimports tool [1]. The following command was run to automatically
format the source code:

	find . -type f -name '*.go' | xargs -I '{}' goimports -w '{}'

Lastly, external dependencies were resolved by running the following
command:

	go mod tidy

[1]: golang.org/x/tools/cmd/goimports

* simplify range loops

The gofmt tool was used to automatically simplify the syntax used
in range loops which only access they key (but not the value) of
the (key, value)-pair. The following command was run to automatically
perform the change:

	find . -type f -name '*.go' | xargs -I '{}' gofmt -s -w '{}'